+++ src/ts/Ads/Managers/OldCampaignRefreshManager.ts
@@ -212,6 +212,8 @@ export class OldCampaignRefreshManager extends RefreshManager {
         const isPromoWithoutProduct = campaign instanceof PromoCampaign && !PurchasingUtilities.isProductAvailable(campaign.getIapProductId());
 
         if (isPromoWithoutProduct) {
+            const productID = (<PromoCampaign>campaign).getIapProductId();
+            this._nativeBridge.Sdk.logWarning(`Promo placement: ${placementId} does not have the corresponding product: ${productID} available`);
             this.onNoFill(placementId);
         } else {
             this.setPlacementReady(placementId, campaign);
@@ -232,23 +234,22 @@ export class OldCampaignRefreshManager extends RefreshManager {
     }
 
     private onError(error: WebViewError | Error, placementIds: string[], diagnosticsType: string, session?: Session) {
-        let errorInternal = error;
         this.invalidateCampaigns(this._needsRefill, placementIds);
 
         if(error instanceof Error) {
-            errorInternal = { 'message': error.message, 'name': error.name, 'stack': error.stack };
+            error = { 'message': error.message, 'name': error.name, 'stack': error.stack };
         }
 
         if(session) {
             SessionDiagnostics.trigger(diagnosticsType, {
-                error: errorInternal
+                error: error
             }, session);
         } else {
             Diagnostics.trigger(diagnosticsType, {
-                error: errorInternal
+                error: error
             });
         }
-        this._nativeBridge.Sdk.logError(JSON.stringify(errorInternal));
+        this._nativeBridge.Sdk.logError(JSON.stringify(error));
 
         const minimumRefreshTimestamp = Date.now() + RefreshManager.ErrorRefillDelay * 1000;
         if(this._refillTimestamp === 0 || this._refillTimestamp > minimumRefreshTimestamp) {