+++ src/ts/Ads/Views/NewVideoOverlay.ts
@@ -6,24 +6,25 @@ import { Localization } from 'Core/Utilities/Localization';
 import { Template } from 'Core/Utilities/Template';
 
 import NewVideoOverlayTemplate from 'html/NewVideoOverlay.html';
+import { ABGroup, ExitSkipIconTest } from 'Core/Models/ABGroup';
 
 export class NewVideoOverlay extends AbstractVideoOverlay implements IPrivacyHandler {
     private _localization: Localization;
 
     private _spinnerEnabled: boolean = false;
 
-    protected _skipEnabled: boolean;
+    private _skipEnabled: boolean;
 
     private _videoDurationEnabled: boolean = false;
-    protected _videoProgress: number;
+    private _videoProgress: number;
 
     private _muteEnabled: boolean = false;
 
     private _debugMessageVisible: boolean = false;
     private _callButtonVisible: boolean = false;
-    protected _callButtonEnabled: boolean = true;
+    private _callButtonEnabled: boolean = true;
 
-    protected _skipButtonElement: HTMLElement;
+    private _skipButtonElement: HTMLElement;
     private _spinnerElement: HTMLElement;
     private _muteButtonElement: HTMLElement;
     private _debugMessageElement: HTMLElement;
@@ -31,15 +32,16 @@ export class NewVideoOverlay extends AbstractVideoOverlay implements IPrivacyHan
     private _timerElement: HTMLElement;
 
     private _fadeTimer: any;
-    protected _areControlsVisible: boolean = false;
+    private _areControlsVisible: boolean = false;
     private _privacy: AbstractPrivacy;
     private _gdprPopupClicked: boolean = false;
     private _showGDPRBanner: boolean = false;
     private _showPrivacyDuringVideo: boolean | undefined;
     private _gameId: string;
     private _country: string | undefined;
+    private _abGroup: ABGroup;
 
-    constructor(nativeBridge: NativeBridge, muted: boolean, language: string, gameId: string, privacy: AbstractPrivacy, showGDPRBanner: boolean, showPrivacyDuringVideo?: boolean, country?: string) {
+    constructor(nativeBridge: NativeBridge, muted: boolean, language: string, gameId: string, privacy: AbstractPrivacy, showGDPRBanner: boolean, abGroup: ABGroup, showPrivacyDuringVideo?: boolean, country?: string) {
         super(nativeBridge, 'new-video-overlay', muted);
 
         this._localization = new Localization(language, 'overlay');
@@ -48,6 +50,7 @@ export class NewVideoOverlay extends AbstractVideoOverlay implements IPrivacyHan
         this._gameId = gameId;
         this._template = new Template(NewVideoOverlayTemplate, this._localization);
         this._country = country;
+        this._abGroup = abGroup;
 
         this._templateData = {
             muted
@@ -136,8 +139,10 @@ export class NewVideoOverlay extends AbstractVideoOverlay implements IPrivacyHan
             }
         }
 
-        if(CustomFeatures.isCheetahGame(this._gameId)) {
+        if (CustomFeatures.isCheetahGame(this._gameId)) {
             this._skipButtonElement.classList.add('close-icon-skip');
+        } else if (ExitSkipIconTest.isValid(this._abGroup)) {
+            this._skipButtonElement.classList.add('exit-icon-skip');
         }
     }
 
@@ -292,7 +297,7 @@ export class NewVideoOverlay extends AbstractVideoOverlay implements IPrivacyHan
         this._handlers.forEach(handler => handler.onOverlayMute(this._muted));
     }
 
-    protected onCallButtonEvent(event: Event): void {
+    private onCallButtonEvent(event: Event): void {
         if (!this._callButtonEnabled) {
             return;
         }
@@ -328,7 +333,7 @@ export class NewVideoOverlay extends AbstractVideoOverlay implements IPrivacyHan
         this._timerElement = <HTMLElement>this._container.querySelector('.timer');
     }
 
-    protected showSkipButton() {
+    private showSkipButton() {
         if (this._skipEnabled) {
             this._skipButtonElement.classList.add('show-skip-button');
         }
@@ -341,7 +346,7 @@ export class NewVideoOverlay extends AbstractVideoOverlay implements IPrivacyHan
         }
     }
 
-    protected showCallButton() {
+    private showCallButton() {
         if (!this._areControlsVisible) {
             return;
         }
@@ -349,7 +354,7 @@ export class NewVideoOverlay extends AbstractVideoOverlay implements IPrivacyHan
         this._callButtonElement.classList.add('show-go-text');
     }
 
-    protected fadeIn() {
+    private fadeIn() {
         this._container.classList.add('fade-in');
         this._areControlsVisible = true;
         setTimeout(() => {