+++ src/ts/XPromo/Models/XPromoCampaign.ts
@@ -1,6 +1,7 @@
 import { Image } from 'Ads/Models/Assets/Image';
 import { Video } from 'Ads/Models/Assets/Video';
 import { Campaign, ICampaign } from 'Ads/Models/Campaign';
+import { Asset } from 'Ads/Models/Assets/Asset';
 
 export enum StoreName {
     APPLE,
@@ -16,8 +17,9 @@ export interface IXPromoCampaign extends ICampaign {
     gameIcon: Image;
     rating: number;
     ratingCount: number;
-    landscapeImage: Image;
-    portraitImage: Image;
+    landscapeImage?: Image;
+    portraitImage?: Image;
+    squareImage?: Image;
     video?: Video;
     streamingVideo?: Video;
     videoPortrait?: Video;
@@ -26,7 +28,6 @@ export interface IXPromoCampaign extends ICampaign {
     clickAttributionUrlFollowsRedirects?: boolean;
     bypassAppSheet: boolean;
     store: StoreName;
-    trackingUrls?: { [eventName: string]: string[] };
     videoEventUrls: { [eventType: string]: string };
 }
 
@@ -40,8 +41,9 @@ export class XPromoCampaign extends Campaign<IXPromoCampaign> {
             gameIcon: ['object'],
             rating: ['number'],
             ratingCount: ['number'],
-            landscapeImage: ['object'],
-            portraitImage: ['object'],
+            landscapeImage: ['object', 'undefined'],
+            portraitImage: ['object', 'undefined'],
+            squareImage: ['object', 'undefined'],
             video: ['object', 'undefined'],
             streamingVideo: ['object', 'undefined'],
             videoPortrait: ['object', 'undefined'],
@@ -50,7 +52,6 @@ export class XPromoCampaign extends Campaign<IXPromoCampaign> {
             clickAttributionUrlFollowsRedirects: ['boolean', 'undefined'],
             bypassAppSheet: ['boolean'],
             store: ['number'],
-            trackingUrls: ['object', 'undefined'],
             videoEventUrls: ['object']
         }, campaign);
     }
@@ -83,12 +84,16 @@ export class XPromoCampaign extends Campaign<IXPromoCampaign> {
         return this.get('ratingCount');
     }
 
-    public getPortrait(): Image {
+    public getPortrait(): Image | undefined {
         return this.get('portraitImage');
     }
 
-    public getLandscape(): Image {
+    public getLandscape(): Image | undefined {
         return this.get('landscapeImage');
+    }
+
+    public getSquare(): Image | undefined {
+        return this.get('squareImage');
     }
 
     public getVideo(): Video | undefined {
@@ -119,14 +124,6 @@ export class XPromoCampaign extends Campaign<IXPromoCampaign> {
         return this.get('bypassAppSheet');
     }
 
-    public getTrackingUrlsForEvent(eventName: string): string[] {
-        const trackingUrls = this.get('trackingUrls');
-        if (trackingUrls) {
-            return (<any>trackingUrls)[eventName] || [];
-        }
-        return [];
-    }
-
     public getVideoEventUrl(eventType: string): string {
         return this.get('videoEventUrls')[eventType];
     }
@@ -140,11 +137,26 @@ export class XPromoCampaign extends Campaign<IXPromoCampaign> {
     }
 
     public getOptionalAssets() {
-        return [
-            this.getGameIcon(),
-            this.getPortrait(),
-            this.getLandscape()
-        ];
+        const assets: Asset[] = [];
+
+        assets.push(this.getGameIcon());
+
+        const square = this.getSquare();
+        if(square) {
+            assets.push(square);
+        }
+
+        const landscape = this.getLandscape();
+        if(landscape) {
+            assets.push(landscape);
+        }
+
+        const portrait = this.getPortrait();
+        if(portrait) {
+            assets.push(portrait);
+        }
+
+        return assets;
     }
 
     public isConnectionNeeded(): boolean {
@@ -156,6 +168,12 @@ export class XPromoCampaign extends Campaign<IXPromoCampaign> {
         const gameIconObject = this.getGameIcon();
         if (gameIconObject) {
             gameIcon = gameIconObject.getDTO();
+        }
+
+        let squareImage: any;
+        const squareImageObject = this.getSquare();
+        if (squareImageObject) {
+            squareImage = squareImageObject.getDTO();
         }
 
         let landscapeImage: any;
@@ -190,6 +208,7 @@ export class XPromoCampaign extends Campaign<IXPromoCampaign> {
             'gameIcon': gameIcon,
             'rating': this.getRating(),
             'ratingCount': this.getRatingCount(),
+            'squareImage': squareImage,
             'landscapeImage': landscapeImage,
             'portraitImage': portraitImage,
             'video': video,