+++ test/Unit/AdTypes/Banner/AdUnits/BannerAdUnitFactoryTest.ts
@@ -1,6 +1,6 @@
-import { IAdUnitParameters } from 'Ads/AdUnits/AbstractAdUnit';
-import { BannerAdUnit } from 'Banners/AdUnits/BannerAdUnit';
 import { BannerAdUnitFactory } from 'Banners/AdUnits/BannerAdUnitFactory';
+import { DisplayHTMLBannerAdUnit } from 'Banners/AdUnits/DisplayHTMLBannerAdUnit';
+import { IBannerAdUnitParameters } from 'Banners/AdUnits/HTMLBannerAdUnit';
 import { BannerCampaign } from 'Banners/Models/BannerCampaign';
 import { assert } from 'chai';
 import { NativeBridge } from 'Core/Native/Bridge/NativeBridge';
@@ -11,19 +11,19 @@ describe('BannerAdUnitFactoryTest', () => {
     describe('when creating an ad unit for a BannerCampaign', () => {
         let campaign: BannerCampaign;
         let nativeBridge: NativeBridge;
-        let parameters: IAdUnitParameters<BannerCampaign>;
+        let parameters: IBannerAdUnitParameters;
 
         beforeEach(() => {
             campaign = sinon.createStubInstance(BannerCampaign);
             nativeBridge = sinon.createStubInstance(NativeBridge);
-            parameters = <IAdUnitParameters<BannerCampaign>>{
+            parameters = <IBannerAdUnitParameters>{
                 campaign
             };
         });
 
         it('should return a banner ad unit', () => {
             const adUnit = BannerAdUnitFactory.createAdUnit(nativeBridge, parameters);
-            assert.instanceOf(adUnit, BannerAdUnit, 'Returned ad unit is not a BannerAdUnit.');
+            assert.instanceOf(adUnit, DisplayHTMLBannerAdUnit, 'Returned ad unit is not a BannerAdUnit.');
         });
     });
 });