+++ CHANGELOG.md
@@ -1,3 +1,67 @@
+# Thu Nov  8 14:32:20 UTC 2018
+
+* Load square asset when constructing backup campaign [#7074](https://github.com/Applifier/unity-ads-webview/pull/7074)
+
+# Thu Nov  8 11:34:51 UTC 2018
+
+* Move trackingUrls to base Campaign model [#7053](https://github.com/Applifier/unity-ads-webview/pull/7053)
+* Fix OrganizationID Description [#6941](https://github.com/Applifier/unity-ads-webview/pull/6941)
+
+# Wed Nov  7 18:30:55 UTC 2018
+
+* AB test for different green shades on Android CTA button [#6891](https://github.com/Applifier/unity-ads-webview/pull/6891)
+
+# Wed Nov  7 13:47:22 UTC 2018
+
+* Use proper keyname for WebPlayer returnValue [#7012](https://github.com/Applifier/unity-ads-webview/pull/7012)
+* Revert TSLint : no-parameter-reassignment [#6897](https://github.com/Applifier/unity-ads-webview/pull/6897)
+* Fix GDPR Banner for AR MRAID [#6943](https://github.com/Applifier/unity-ads-webview/pull/6943)
+* Log Promo error whenever no creatives detected [#6803](https://github.com/Applifier/unity-ads-webview/pull/6803)
+* Respect iPhone x safe areas for Admob [#6989](https://github.com/Applifier/unity-ads-webview/pull/6989)
+* Refactor Banner Ad Unit [#6889](https://github.com/Applifier/unity-ads-webview/pull/6889)
+
+# Tue Nov  6 21:05:41 UTC 2018
+
+* Exit skip button test [#6991](https://github.com/Applifier/unity-ads-webview/pull/6991)
+* Revert skippable rewarded ads test [#6990](https://github.com/Applifier/unity-ads-webview/pull/6990)
+* SafeDK hacks [#6992](https://github.com/Applifier/unity-ads-webview/pull/6992)
+* Add a global afterEach which restores sinon default sandbox [#6968](https://github.com/Applifier/unity-ads-webview/pull/6968)
+
+# Tue Nov  6 11:53:35 UTC 2018
+
+* TSLint no-relative-imports [#6888](https://github.com/Applifier/unity-ads-webview/pull/6888)
+* Remove feature flag test for vast3 error tracking [#6895](https://github.com/Applifier/unity-ads-webview/pull/6895)
+* Update AR styling for iPhoneX and Info button [#6837](https://github.com/Applifier/unity-ads-webview/pull/6837)
+* MRAID Views Refactor [#6775](https://github.com/Applifier/unity-ads-webview/pull/6775)
+* Resize MRAID Views [#6782](https://github.com/Applifier/unity-ads-webview/pull/6782)
+
+# Mon Nov  5 13:44:02 UTC 2018
+
+* Block file URLs from all Request and Cache API invocations [#6948](https://github.com/Applifier/unity-ads-webview/pull/6948)
+
+# Thu Nov  1 20:15:06 UTC 2018
+
+* Use X as skip icon experiment [#6920](https://github.com/Applifier/unity-ads-webview/pull/6920)
+* Remove early CTA button experiment [#6918](https://github.com/Applifier/unity-ads-webview/pull/6918)
+
+# Wed Oct 31 23:02:03 UTC 2018
+
+* Refactor AdUnitFactory [#6834](https://github.com/Applifier/unity-ads-webview/pull/6834)
+* Remove expired ab group test for click delay [#6861](https://github.com/Applifier/unity-ads-webview/pull/6861)
+* Fix error diagnostics for Content Types [#6862](https://github.com/Applifier/unity-ads-webview/pull/6862)
+* Update Error handling for Videos over 40 seconds [#6886](https://github.com/Applifier/unity-ads-webview/pull/6886)
+* Vast media selector abTest using auction feature flag [#6859](https://github.com/Applifier/unity-ads-webview/pull/6859)
+* Encode VAST Wrapper URI with params [#6858](https://github.com/Applifier/unity-ads-webview/pull/6858)
+* Refactor BannerAdUnitParametersFactory to remove unused parameters [#6860](https://github.com/Applifier/unity-ads-webview/pull/6860)
+
+# Tue Oct 30 17:52:40 UTC 2018
+
+* Fix Privacy layout on iPhone X series [#6863](https://github.com/Applifier/unity-ads-webview/pull/6863)
+* iPhone XR media queries [#6864](https://github.com/Applifier/unity-ads-webview/pull/6864)
+* Square end screen asset support [#6838](https://github.com/Applifier/unity-ads-webview/pull/6838)
+* Experiment: Skippable Rewarded Ads [#6806](https://github.com/Applifier/unity-ads-webview/pull/6806)
+* Log Promotions without product [#6780](https://github.com/Applifier/unity-ads-webview/pull/6780)
+
 # Mon Oct 29 21:07:45 UTC 2018
 
 * Refactor Backend to non-static [#6835](https://github.com/Applifier/unity-ads-webview/pull/6835)